Building class

+++ src/exercises_session_4/DivideTenBy.java
@@ -8,12 +8,12 @@ public class DivideTenBy {
         int[]  myArray = {2,6,8,1,9,0,10,23,7,5,3};
 
         for (int j : myArray) {
-            int divided = 0;
+            double divided = 0;
 
             System.out.println("Number: " + j);
             if(j%2 == 0) {
                 try {
-                    divided = 10/j;
+                    divided = (double) 10/ (int) j;
                 } catch (ArithmeticException ae) {
                     System.out.println("You divided by zero!");
                     continue;
